Alterar senha funcionando

+++ MyIce-BackEnd/src/main/java/ecommerce/Repository/ClienteRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Long>  {
+public interface ClienteRepository extends JpaRepository<Cliente, Integer>  {
     public Cliente findByEmail(String email);
     public Cliente findByIdClienteAndPassword(Integer idCliente, String password);
